ref: NewVRRPmulticastPacket()
